requirements.txt
certifi==2022.12.7
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
cycler==0.11.0
Detectron==0.0.0
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
h5py==3.8.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.7.0
joblib==1.3.1
kiwisolver==1.4.4
matplotlib==3.5.3
nltk==3.8.1
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
pip==22.3.1
protobuf==4.23.4
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.2
setuptools==65.6.3
six==1.16.0
tensorboardX==2.6.2
tokenizers==0.13.3
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.65.1
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.4
wheel==0.38.4
zipp==3.15.0
